Initial changes to support Tcl/Tk 8.5.
Note also requires updates to itcl/itk 3.4 and BLT 3.0a and at this commit
requires that support for the TK_CONFIG_OPTION_SPECIFIED flag in
Tk_ConfigureWidget is restored (this flag has been removed for thread safety
reasons).  A move to the Tk_SetOptions API is needed to remove this
requirement (these sourcesets are available from the JAC git repository)

Most of the changes are just to use modern fonts: TkDefaultFont TkFixedFont
and TkHeadingFont. Alpha and dec symbols are now selected using unicode.

Other changes include the removal of my canvas printing support. Scoping
changes in Tk mean this is no longer possible and has to be done as a
patch of Tk (the support is optional).

Finally creating big lists with Tcl_AppendElement is really slow now, so
some changes to the object interface have been made (catalogue query command).